# docs/swagger.yaml
basePath: /api
definitions:
  main.Area:
    description: A specific area within a room where hedgehogs are housed
    properties:
      created_at:
        example: "2024-01-15T10:30:00Z"
        format: date-time
        type: string
      hedgehogs:
        items:
          $ref: '#/definitions/main.Hedgehog'
        type: array
      height:
        example: 80
        minimum: 1
        type: number
      id:
        example: 1
        type: integer
      max_capacity:
        example: 2
        minimum: 1
        type: integer
      name:
        example: Gabbia 1
        type: string
      room:
        $ref: '#/definitions/main.Room'
      room_id:
        example: 1
        type: integer
      updated_at:
        example: "2024-01-15T10:30:00Z"
        format: date-time
        type: string
      width:
        example: 100
        minimum: 1
        type: number
      x:
        example: 10
        minimum: 0
        type: number
      "y":
        example: 20
        minimum: 0
        type: number
    type: object
...
  main.LoginRequest:
    properties:
      password:
        example: admin123
        type: string
      username:
        example: admin
        type: string
    required:
    - password
    - username
    type: object
...
  main.TherapyAnalysis:
    properties:
      alert:
        type: boolean
      alert_reason:
        type: string
      days_active:
        type: integer
      days_until_end:
        type: integer
      end_date:
        type: string
      hedgehog_id:
        type: integer
      hedgehog_name:
        type: string
      start_date:
        type: string
      status:
        type: string
      therapy_id:
        type: integer
      therapy_name:
        type: string
    type: object
  main.TokenResponse:
    properties:
      expires_at:
        example: 1640995200
        type: integer
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
  main.WeightAnalysis:
    properties:
      alert:
        type: boolean
      alert_reason:
        type: string
      current_weight:
        type: number
      days_since_weigh:
        type: integer
      hedgehog_id:
        type: integer
      hedgehog_name:
        type: string
      last_weigh_date:
        type: string
      previous_weight:
        type: number
      trend:
        description: '"improving", "stable", "declining", "critical"'
        type: string
      weight_change:
        type: number
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for managing hedgehog rescue center operations
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: La Ninna - Hedgehog Management API
  version: "1.0"
paths:
  /analysis/therapy:
    get:
      consumes:
      - application/json
      description: Get analysis of active therapies with alerts for expiring or overdue
        therapies
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.TherapyAnalysis'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get therapy analysis
      tags:
      - Analysis
  /analysis/weight:
    get:
      consumes:
      - application/json
      description: Get analysis of hedgehog weight trends and alerts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.WeightAnalysis'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get weight analysis
      tags:
      - Analysis
  /areas:
    get:
      consumes:
      - application/json
      description: Get list of all areas with their rooms and hedgehogs
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Area'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all areas
      tags:
      - Areas
    post:
      consumes:
      - application/json
      description: Create a new area record
      parameters:
      - description: Area data
        in: body
        name: area
        required: true
        schema:
          $ref: '#/definitions/main.Area'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.Area'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create new area
      tags:
      - Areas
  /areas/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an area by its ID
      parameters:
      - description: Area ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete area
      tags:
      - Areas
    put:
      consumes:
      - application/json
      description: Update an existing area's information
      parameters:
      - description: Area ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated area data
        in: body
        name: area
        required: true
        schema:
          $ref: '#/definitions/main.Area'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Area'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update area
      tags:
      - Areas
...
  /hedgehogs:
    get:
      consumes:
      - application/json
      description: Get list of all hedgehogs with their areas and relationships
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Hedgehog'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all hedgehogs
      tags:
      - Hedgehogs
    post:
      consumes:
      - application/json
      description: Create a new hedgehog record
      parameters:
      - description: Hedgehog data
        in: body
        name: hedgehog
        required: true
        schema:
          $ref: '#/definitions/main.Hedgehog'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.Hedgehog'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create new hedgehog
      tags:
      - Hedgehogs
  /hedgehogs/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a hedgehog by its ID
      parameters:
      - description: Hedgehog ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete hedgehog
      tags:
      - Hedgehogs
    get:
      consumes:
      - application/json
      description: Get a single hedgehog by its ID with related data
      parameters:
      - description: Hedgehog ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Hedgehog'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get hedgehog by ID
      tags:
      - Hedgehogs
    put:
      consumes:
      - application/json
      description: Update an existing hedgehog's information
      parameters:
      - description: Hedgehog ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated hedgehog data
        in: body
        name: hedgehog
        required: true
        schema:
          $ref: '#/definitions/main.Hedgehog'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Hedgehog'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update hedgehog
      tags:
      - Hedgehogs
  /login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT token
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/main.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.TokenResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User login
      tags:
      - Authentication
  /notification-settings:
    get:
      consumes:
      - application/json
      description: Get the current notification settings for the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.NotificationSettings'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get notification settings
      tags:
      - Settings
    put:
      consumes:
      - application/json
      description: Update the notification settings for the system
      parameters:
      - description: Notification settings
        in: body
        name: settings
        required: true
        schema:
          $ref: '#/definitions/main.NotificationSettings'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.NotificationSettings'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update notification settings
      tags:
      - Settings
  /notifications:
    get:
      consumes:
      - application/json
      description: Get list of notifications with optional filters
      parameters:
      - description: Filter by unread status
        in: query
        name: unread
        type: boolean
      - description: Filter by priority
        in: query
        name: priority
        type: string
      - description: Filter by notification type
        in: query
        name: type
        type: string
      - default: 50
        description: Limit results
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Notification'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get notifications
      tags:
      - Notifications
  /notifications/{id}:
    delete:
      consumes:
      - application/json
      description: Dismiss a notification by its ID
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Dismiss notification
      tags:
      - Notifications
  /notifications/{id}/read:
    put:
      consumes:
      - application/json
      description: Mark a notification as read by its ID
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Notification'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Mark notification as read
      tags:
      - Notifications
  /notifications/stats:
    get:
      consumes:
      - application/json
      description: Get statistics about notifications (counts by status, priority,
        and type)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get notification statistics
      tags:
      - Notifications
  /rooms:
    get:
      consumes:
      - application/json
      description: Get list of all rooms with their areas and hedgehogs
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Room'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all rooms
      tags:
      - Rooms
    post:
      consumes:
      - application/json
      description: Create a new room record
      parameters:
      - description: Room data
        in: body
        name: room
        required: true
        schema:
          $ref: '#/definitions/main.Room'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.Room'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create new room
      tags:
      - Rooms
  /rooms/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a room by its ID
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete room
      tags:
      - Rooms
    get:
      consumes:
      - application/json
      description: Get a single room by its ID with related areas and hedgehogs
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Room'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get room by ID
      tags:
      - Rooms
    put:
      consumes:
      - application/json
      description: Update an existing room's information
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated room data
        in: body
        name: room
        required: true
        schema:
          $ref: '#/definitions/main.Room'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Room'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update room
      tags:
      - Rooms
  /therapies:
    get:
      consumes:
      - application/json
      description: Get list of all therapies with optional hedgehog_id filter
      parameters:
      - description: Filter by hedgehog ID
        in: query
        name: hedgehog_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Therapy'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all therapies
      tags:
      - Therapies
    post:
      consumes:
      - application/json
      description: Create a new therapy record for a hedgehog
      parameters:
      - description: Therapy data
        in: body
        name: therapy
        required: true
        schema:
          $ref: '#/definitions/main.Therapy'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.Therapy'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create new therapy
      tags:
      - Therapies
  /therapies/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a therapy by its ID
      parameters:
      - description: Therapy ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete therapy
      tags:
      - Therapies
    put:
      consumes:
      - application/json
      description: Update an existing therapy's information
      parameters:
      - description: Therapy ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated therapy data
        in: body
        name: therapy
        required: true
        schema:
          $ref: '#/definitions/main.Therapy'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Therapy'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update therapy
      tags:
      - Therapies
  /weight-records:
    get:
      consumes:
      - application/json
      description: Get weight records with optional hedgehog_id filter
      parameters:
      - description: Filter by hedgehog ID
        in: query
        name: hedgehog_id
        type: integer
      - default: 100
        description: Limit results
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.WeightRecord'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get weight records
      tags:
      - Weight Records
    post:
      consumes:
      - application/json
      description: Add new weight record for a hedgehog
      parameters:
      - description: Weight record data
        in: body
        name: weight_record
        required: true
        schema:
          $ref: '#/definitions/main.WeightRecord'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.WeightRecord'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create weight record
      tags:
      - Weight Records
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"